Initial commit
Unfortunately, the git repository history of this was lost.
The source remains entirely intact, though, and I'm genuinely intrigued
by the amount of code that I wrote, the solutions to problems I came up
with, and the amount of evolution that the project went through from
it's original early versions. I'm reviving this as a research project
for myself to rehabilitate the code and get it in decent condition
again.